refactor(databases): use route-driven branch tabs
