+++ apps/app/src/app/(workspace)/projects/[id]/environments/[envId]/databases/[databaseId]/branches/[targetId]/settings/page.tsx
@@ -1,0 +1,3 @@
+export default function DatabaseBranchSettingsRoute() {
+  return null;
+}
