+++ apps/app/src/app/(workspace)/projects/[id]/environments/[envId]/databases/[databaseId]/branches/[targetId]/sql/page.tsx
@@ -1,0 +1,3 @@
+export default function DatabaseBranchSqlRoute() {
+  return null;
+}
